backend/requirements.txt
supabase==2.6.0
python-dotenv==1.0.1